[Add] 장비 착용/해제 시 플레이어 스탯 재계산 연동
- PlayerStats: 체력 외 공격력/방어력/이동속도 필드 추가 (base값 + 최종값 구조)

- Inventory.SumModifiers(StatType): 장착된 3개 슬롯의 StatModifiers 합산

- Inventory ↔ PlayerStats는 VoidEventChannel(장착/해제)로 간접 연결

+++ Assets/Scripts/ItemData.cs
@@ -8,7 +8,7 @@ public enum EquipmentSlotType { Weapon, Armor, Accessory }
 public struct StatModifier
 {
     public StatType Type;
-    public float Value;
+    public int Value;
 }
 
 [CreateAssetMenu(menuName ="Data/ItemData")]
